+++ CHANGELOG.md
@@ -1,7 +1,13 @@
 # Pulse 3.x
 
-- [macOS] Fix an issue where changing sort descriptors in a table view would affect other display modes
-- [macOS] Fix an issue where on first selection in a text view or search results list, the main panel would reload 
+### PulseUI on macOS
+
+- Fix an issue where changing sort descriptors in a table view would affect other display modes
+- Fix an issue where on first selection in a text view or search results list, the main panel would reload 
+
+### Pulse
+
+- Remove `NetworkLoggerInsights`
 
 ## Pulse 3.4.3