+++ CHANGELOG.md
@@ -1,9 +1,10 @@
 # Pulse 3.x
 
-### PulseUI on macOS
+### PulseUI
 
-- Fix an issue where changing sort descriptors in a table view would affect other display modes
-- Fix an issue where on first selection in a text view or search results list, the main panel would reload 
+- [macOS] Fix an issue where changing sort descriptors in a table view would affect other display modes
+- [macOS] Fix an issue where on first selection in a text view or search results list, the main panel would reload 
+- The insights view now shows info for the currently selected tasks
 
 ### Pulse